Fix cron jobs not running when bind mounted on Linux
When bind mounting the file on macOS, Docker automatically changes the
permissions so it was working as expected. On Linux, however, bind
mounting keeps the file permissions as-is. If the file isn't owned by
root with 600 permissions, crond just won't run any of the jobs and
doesn't log any errors.

In the future I might have the config files be mounted in a separate
directory and then just copy them to the correct spot, but I don't think
that it's necessary right now.